Small changes per PR comments
