+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/MavenAppEngineFacetUtil.java
@@ -1,24 +1,14 @@
 package com.google.cloud.tools.eclipse.appengine.facets;
 
-import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
-import org.apache.maven.artifact.Artifact;
 import org.apache.maven.model.Dependency;
-import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.m2e.core.MavenPlugin;
-
-import com.google.common.base.Objects;
+import com.google.cloud.tools.eclipse.util.MavenUtils;
 
 public class MavenAppEngineFacetUtil {
-  private static final Logger LOGGER = Logger.getLogger(MavenAppEngineFacetUtil.class.getName());
-
   /**
    * Returns a list of all the App Engine dependencies that should exist in the pom.xml
    * of a maven project that has the App Engine facet installed
@@ -71,9 +61,9 @@ public class MavenAppEngineFacetUtil {
    * @return a map where the keys and values are the property fields and values respectively
    */
   public static Map<String, String> getAppEnginePomProperties(IProgressMonitor monitor) {
-    String appengineArtifactVersion = resolveLatestReleasedArtifact(monitor,
+    String appengineArtifactVersion = MavenUtils.resolveLatestReleasedArtifact(monitor,
         "com.google.appengine", "appengine-api-1.0-sdk", "jar", AppEngineStandardFacet.DEFAULT_APPENGINE_SDK_VERSION);
-    String gcloudArtifactVersion = resolveLatestReleasedArtifact(monitor,
+    String gcloudArtifactVersion = MavenUtils.resolveLatestReleasedArtifact(monitor,
         "com.google.appengine", "gcloud-maven-plugin", "maven-plugin", AppEngineStandardFacet.DEFAULT_GCLOUD_PLUGIN_VERSION);
 
     Map<String, String> allProperties = new HashMap<String, String>();
@@ -82,55 +72,6 @@ public class MavenAppEngineFacetUtil {
     allProperties.put("appengine.version", appengineArtifactVersion);
     allProperties.put("gcloud.plugin.version", gcloudArtifactVersion);
     return allProperties;
-  }
-
-  /**
-   * Returns true if the group IDs and artifact IDs of <code>dependency1</code> and
-   * <@code>dependency2</@code> are equal. Returns false otherwise.
-   */
-  // visible for testing
-  public static boolean areDependenciesEqual(Dependency dependency1, Dependency dependency2) {
-    if ((dependency1 == null) || (dependency2 == null)) {
-      return false;
-    }
-
-    if (!Objects.equal(dependency1.getGroupId(), dependency2.getGroupId())||
-        !Objects.equal(dependency1.getGroupId(), dependency2.getGroupId())) {
-      return false;
-    }
-
-    return true;
-  }
-
-  /**
-   * Returns true if a dependency with the same group ID and artifact ID as <code>targetDependency</code>
-   * exists in <code>dependencies</code>. Returns false otherwise.
-   */
-  public static boolean doesListContainDependency(List<Dependency> dependencies, Dependency targetDependency) {
-    if((dependencies == null) || (targetDependency == null)) {
-      return false;
-    }
-
-    for (Dependency dependency : dependencies) {
-      if (areDependenciesEqual(dependency, targetDependency)) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  public static String resolveLatestReleasedArtifact(IProgressMonitor monitor, String groupId,
-      String artifactId, String type, String defaultVersion) {
-    try {
-      Artifact artifact = MavenPlugin.getMaven().resolve(groupId, artifactId, "LATEST", type,
-          null /* classifier */, null /* artifactRepositories */, monitor);
-      return artifact.getVersion();
-    } catch (CoreException ex) {
-      LOGGER.log(Level.WARNING,
-          MessageFormat.format("Unable to resolve artifact {0}:{1}", groupId, artifactId), ex);
-      return defaultVersion;
-    }
   }
 
 }
