+++ plugins/com.google.cloud.tools.eclipse.util/META-INF/MANIFEST.MF
@@ -20,3 +20,5 @@ Import-Package: com.google.common.base;version="15.0.0",
 Bundle-ClassPath: lib/freemarker.jar,
  .,
  templates/appengine/
+Require-Bundle: org.eclipse.m2e.maven.runtime,
+ org.eclipse.m2e.core;bundle-version="1.6.2"
